Add files via upload
                                                                                                                               Abstract
Volatility forecasting is a vital task for asset valuation and risk management for investors and financial intermediaries. Various volatility forecasting models have been introduced and practiced over the period. However, the demand for better prediction models is always there as the financial markets update itself on a continuous basis. The purpose of this thesis is to examine the volatility forecasting performance of different Tree Based Regression models with GARCH models on NIFTY50 stock index. The GARCH models examined in this thesis are SGARCH, EGARCH and GJR-GARCH. And the Tree Based Regression models examined are Bagging, Random Forest and Gradient Boosting. The statistical loss functions used for evaluating the performance of the models are MSE and RMSE. The result of this thesis is that Tree Based Regression models have better prediction accuracy compared to GARCH models. Within Tree Based Regression Models Gradient Boosting method outperformed all other models with least MSE and RMSE. This is because of the fact that Gradient Boosting approach utilizes sequential training by constructing an ensemble of shallow and weak consecutive trees, each of which learns from and improves upon itself.